return tasks from actions
